Split the angular code into separate files, based on the john papa guide recommendations. Also configured simple sourcemaps for webpack.
