+++ js/cyclone.module.js
@@ -1,0 +1,15 @@
+angular.module("cycloneApp", ["firebase", 'ngMaterial', 'ngRoute']);
+
+// Routing
+angular.module("cycloneApp").config(function($routeProvider, $locationProvider) {
+    $routeProvider
+        .when('/:type', {templateUrl: 'index.html', controller: 'TimeCtrl'})
+        .when('/:type/:year/:month/:day*', {templateUrl: 'index.html', controller: 'TimeCtrl'})
+        .when('/:type/:weekNumber/:weekDay*', {templateUrl: 'index.html', controller: 'TimeCtrl'})
+        .otherwise({redirectTo: '/today'});
+
+    $locationProvider.html5Mode({
+        enabled: false,
+        requireBase: false
+    });
+});
